Add more verification.
Cleanup and extend IR verification to more closely match XLS's.
